Últimos parches...

+++ src/app/components/users/user-list/user-list.component.ts
@@ -11,16 +11,16 @@ import { UserService } from 'src/app/services/user.service';
 })
 export class UserListComponent implements OnInit {
   /**Uusario logeado? */
-  userLoged:User | undefined = this.authService.getCurrentUser();
+  userLoged: User | undefined = this.authService.getCurrentUser();
 
   /**Varible donde vamos a guardar todos los usuarios */
-  users: User [] | undefined = [];
+  users: User[] | undefined = [];
 
   constructor(
-    private authService:AuthService,
-    private userService:UserService,
-    private router:Router
-  ){}
+    private authService: AuthService,
+    private userService: UserService,
+    private router: Router
+  ) { }
 
   ngOnInit(): void {
     this.userService.getUsersHttp().subscribe({
@@ -34,22 +34,46 @@ export class UserListComponent implements OnInit {
 
     /**El guard casero */
     setTimeout(() => {
-      if(this.userLoged?.rol === 1){
+      if (this.userLoged?.rol === 1) {
         this.router.navigate(["home"]);
       }
-    },5000)
+    }, 5000)
   }
 
-  deleteUser(userId:number){
-    this.userService.deleteUserHttp(userId).subscribe({
-      next: () => {
-        if(confirm("Are you sure?")){
+  deleteUser(userId: number) {
+    const ok = confirm(`Seguro que quiere eliminar el usuario con la id ${userId} ?`);
+
+    if (ok) {
+      this.userService.deleteUserHttp(userId).subscribe({
+        next: () => {
           alert(`El usuario con la id ${userId} fue eliminado`);
           window.location.reload();
-        }else{
-          return
         }
+      })
+    } else {
+      return
+    }
+
+  }
+
+  addAdmin(user: User) {
+    let u = user;
+    if (u.rol !== 0) {
+      const ok = confirm(`Seguro que quiere hacer administrador al ususario con la id ${user.id} ?`);
+      if (ok) {
+        u.rol = 0;
+        this.userService.putUserHttp(u).subscribe({
+          next: () => {
+            alert("Adminsitrador agregado correctamente...");
+            window.location.reload();
+          }
+        });
+      } else {
+        return
       }
-    })
+    } else {
+      alert("El usuario ya es administrador...");
+      return
+    }
   }
 }
